[beta] :memo: tryna catch errors 2

+++ app/src/main/java/com/shub39/rush/domain/interfaces/OtherPreferences.kt
@@ -12,8 +12,8 @@ interface OtherPreferences {
     fun getAppThemePrefFlow(): Flow<AppTheme>
     suspend fun updateAppThemePref(pref: AppTheme)
 
-    fun getSeedColorFlow(): Flow<Long>
-    suspend fun updateSeedColor(newCardContent: Long)
+    fun getSeedColorFlow(): Flow<Int>
+    suspend fun updateSeedColor(newCardContent: Int)
 
     fun getAmoledPrefFlow(): Flow<Boolean>
     suspend fun updateAmoledPref(amoled: Boolean)
